Make Protocol functional on Windows.
The problem was that iterator was always returning null,
and hence the check in FileProtocolsDB.load() for at least
one entry present in the file was always succeding.

Hence, on Windows, we were always using NetDBParser with empty
content, making it all completely useless.

P.S. With this change, Protocol is useful on Windows, it uses
hardcoded pure-java values though. The next step is to enable
proper native support on Windows.